Dashboard/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
idna==3.10
joblib==1.4.2
mysql-connector-python==9.3.0
numpy==2.2.5
pandas==2.2.3
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-jose==3.4.0
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
